Pull out helper functions from shader_wrapper.py
